@@ -1,17 +1,16 @@
 [metadata]
 name = wmm2015
-version = 1.0.2
+version = 1.1.0
 author = Michael Hirsch, Ph.D.
 url = https://github.com/space-physics/wmm2015
 description = WMM2015 geomagnetic model with simple object-oriented Python interface
 keywords =
   geomagnetic
 classifiers =
-  Development Status :: 4 - Beta
+  Development Status :: 5 - Production/Stable
   Environment :: Console
   Intended Audience :: Science/Research
   Operating System :: OS Independent
-  Programming Language :: Python :: 3.6
   Programming Language :: Python :: 3.7
   Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
@@ -22,7 +21,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >= 3.6
+python_requires = >= 3.7
 include_package_data = True
 zip_safe = False
 packages = find:
